Fix a very rare and major issue that took place if someone killtarget'ed a func_bobbing on a map (not something you commonly do but I found a way to fix it). Issue description:
Make a map and place a func_bobbing on it, a trigger_multiple that killtarget's the func_bobbing, and the hook gun pickup somewhere. Get the hook gun in that map and shoot at the walls with it (make sure the map is large enough so the hook flies for at least one second before hitting a wall). The hook will behave normally. Now trigger the trigger_multiple and watch it make the func_bobbing disappear like it should. However, from this point on, shooting the hook gun will cause the hook to change direction in the air and go toward origin '0 0 0' half of the time you shoot it.

This seems to be because the func_bobbing's controller (that moves it up and down) attaches to the hook projectile if the func_bobbing is killed, for an unknown reason (likely happens with other projectiles too). So I added a check that makes it only change the velocity of func_bobbing as long as its classname corresponds. That way, if someone makes a map where they want to killtarget a func_bobbing for whatever reason, that won't brake projectiles.