Other improvements to the multiple centerprints system
Server now uses Send_CSQC_Centerprint_Generic instead of centerprint_builtin to send messages, so that it's possible to send more parameters (id, duration and countdown_num) without adding them to the msg (like the id before). If duration is not set, then the default duration will be used, that is hud_panel_centerprint_time.

In the countdown case, the server can send only the first msg of a series with the same id and the client can continue the countdown by its own; the next countdown messages can be sent by the server anyway, for example to change the countdown msg or for whatever reason. Countdown can be stopped by the server at any time by sending an empty msg (always same id).

The centerprint function in the client now is a wrapper for centerprint_generic, which actually handles the new parameters and add the msgs to the list.

Changes on minstagib_ammocheck are neccessary to correctly stop the out of ammo countdown.

Still other work to do anyway...