Balanced Raptor and Racer a bit
The Raptor was a bit strong and was a general purpose weapon against
everything. It is now a lot weaker against other Raptors (no aim helper
anymore) and a bit weaker against racers (a bit harder to hit) while
still being the weapon of choice against spiderbots (if you manager to
get it up). To prevent the racer from taking out the weaker raptors too
easily the locking time was set up to make it harder to lock on them.
Locking on a raptor which did not notice you is easily possible. But
having a duel with a raptor is a lot harder/interesting.

All in all this works out nicely balance wise (tested for a couple of
weeks on MoN Vehicle Server) and allows far more interesting fights in
air.