Rename the 'find' command to 'findent'
The find command is reserved as an alias of apropos in FTEQW,
Quakespasm, and Source, and assumedly other engines.

In the particular case of FTEQW, compatibility with the engine is
broken and the command is hijacked by FTEQW, making the Xonotic use
of the command impossible. We should try to maintain compatibility
with at least FTEQW as much as possible, on the principle of not
being restricted to a single engine.

As a side note, this functionality ought to become part of the engine
itself.