strafehud: fix and refactor jumpheight feature
The height displayed was always too small, as one or two frames
of each jump were not included (especially the first frame,
when z velocity is greatest).

It now shows the exact height predicted by the formula in physicsX.cfg
and should be correct for all physics configurations.

Also, simplified the logic and removed 6 globals.