Implement automatic per-map min & max player limits
Enabled by g_warmup -1 and g_maxplayers -1 respectively.

Map settings are loaded from .sizes files
and are rounded to a multiple of the number of teams.
At the midpoint, min players is rounded down and max players is rounded up.
Neither can exceed engine maxplayers which is also rounded down if necessary.

g_warmup -1 means stay in warmup until enough players have joined,
then switch to g_warmup_limit and wait for ready players.
"Enough" has a lower limit of 2 or 2 * number of teams,
so this can be useful on maps with no set minimum.