Fix bug where ReadyRestart could cause teamed items to all spawn together
For weapons this was subtle: ghost item(s) would appear as well as
the item selected to spawn.

For powerups they would all spawn at once, and/or previously scheduled
spawns were not cancelled.

Fixing this requires leaving the first of the teamed items marked so
Item_FindTeam() will do its job again at ReadyRestart.  This requires
changing the abused effect field to a bitmask other than EF_NODRAW so
that DP will draw the ghost for the marked item.

Some minor cleanups to Item_FindTeam() are included.