Refactor Nagger_SendEntity and Net_Handle_ENT_CLIENT_NAGGER
This will no longer network ready status outside of warmup_stage in the
case that some clients manage to have ready status then.  This change is
made possible by the removal of sv_ready_restart in
https://gitlab.com/xonotic/xonotic-data.pk3dir/-/merge_requests/940

It also simplifies the code and no longer networks the ready status of
some SVQC entities that are not clients in the case that maxclients is
not a multiple of 8.