Clip velocity to correct plane when trying to step up in a corner
With sv_gameplayfix_stepmultipletimes 1, the wrong plane normal was often
used to modify velocity when trying to step up in a corner, making it
difficult or impossible to step up with certain camera angles and keys.

Optimised stepping up to skip unnecessary traceboxes.

Adjusted push down to only push down as far as the step up could go,
consistent with Q3.  Previously we were cancelling a bit of the player's
positive z velocity in rare cases.

Fixes #2818