pm: implement Q3 and CPMA step up behaviours, improve default
Allows stepping up over tiny ramps that are likely brush misalignments
(when the player first collides with them) in all configs as it's needed
to support several Q3/QL maps (not just defrag ones). This would also be
more forgiving to mappers using blender to create some geometry.
A simple hack is used to make this frametime-independent.

Limits distance for more consistent jump placement and to prevent
missing the jump opportunity entirely when the step's top surface is
narrow and/or speed is very high.