Greatly simplify FlyMove velocity clipping
This is made possible by tracebox.  The previous approach is from Quake
and is flawed such that the triple plane intersection case that sets
velocity to '0 0 0' can be hit even when the player is only in collision
with 1 or 2 planes.  Q3 and DOOM 3 have a more correct (but complex)
approach which I tested but found to be unnecessary (Q3 and DOOM 3 don't
have tracebox).

Fixes #2821

Fixes some cases of sticking when surfing on slopes, helping with #2237