Upgrade the FlyMove velocity clipping logic
This is a port of the Quake 3 / DOOM 3 approach,
which avoids the flaws of the Quake approach (used previously),
with some simplifications made possible by tracebox.

In the Quake approach, the triple plane intersection case that sets
velocity to '0 0 0' can be hit even when the player is only in collision
with 1 or 2 planes.

Fixes #2821

Fixes some cases of sticking when surfing on slopes, helping with #2237