pm: upgrade the core velocity clipping logic to id tech 3
This is the Q3 / DOOM 3 method, which avoids the flaws of the Quake
method used previously.

In the Quake method the triple plane intersection case that sets
velocity to '0 0 0' can be hit even when the player is only in collision
with 1 or 2 planes.

Also the Quake "stop dead to avoid tiny occilations" case can be hit
when the player is trying to jump in a corner formed by two angled
(relative to the map grid) brush faces.

Fixes #2821

Fixes some cases of sticking when surfing on slopes, helping with #2237

---

Updates _Movetype_ClipVelocity to fix precision and logic issues

Removes code inherited from Quake that deletes small amounts of velocity
when direction of travel is almost aligned to an axis. It's incorrect to
change direction when it's not caused by any collision, application of
force, or player input.

Adds the overbounce/OVERCLIP value from Q3 and DOOM 3 to fix some
corner cases where the player can get wedged while still having velocity.

Adds the ob 1 optimisation from DOOM 3.