Remove legacy Quake bbox expansion: projectiles
All the "zero size" projectiles now need a size so they can
trigger a warpzone's touch func. Prevously DP gave most of them
mins '-1 -1 -1' maxs '1 1 1' for the purpose of those tests
(but not all tests).

Because of that these projectiles now need a different way to avoid
colliding with each other. Comparison of the `clipgroup` field is used
as DP_RM_CLIPGROUP will make this more optimal for performance and is
also implemented in QC here for compatibility with old DP versions.