Disable keepaway ball pickup delay
The delay was frustrating when getting a bckill in midair but being
unable to pick up the ball despite having travelled through it.

We still need to prevent instant pickup by the previous owner so the
ball can be dropped manually. Sharing a field from KH for this with the
same 0.5s delay as used when throwing a weapon.