Remove legacy Quake bbox expansion: items (and buffs)
Adds a Small 48x48x48 item bbox which is used for 5 and 25 health and
armor items. This is the biggest horizontal size that fits in the 25h
alcove on finalrage.

Changes powerup, buff and mega pickups to all have the same height.
This will make it cheaper to network item bboxes to CSQC.