items: only draw the removal effect when appropriate
Sends an extra update that triggers the effect prior to removal.
Currently only used for loot items when picked up or touching void.

Fixes #2849
without reverting to the old way of sending the effect in SVQC which
caused every player to receive it even if they couldn't see the item,
and without the expense of additional culling.

Changes the pipeline hash because g_powerups_drop_ondeath is enabled by
default and bots remove an item from their goals immediately if it's
deleted, but if it still exists (for 1-2 frames in this case) they only
remove it if it's in their PVS.