triggers: invert logic of warpzone_isboxy and rename to trigger_exact
This change should prevent obscure bugs when touching triggers while
moving (race-free triggering), which requires knowing if the trigger
uses a bbox or a model: it's easy to identify all triggers that use a
model (by their use of EXACTTRIGGERS_INIT and EXACTTRIGGERS_TOUCH) or
the functions they expand to, but hard to identify all triggers that use
a bbox.

It may also help prevent performance loss because now the expensive code
path is opt-in instead of opt-out.

Moved declaration and renamed because this isn't specific to warpzones.