Remove legacy Quake bbox expansion: CTF
This adds a feature where if the flag is dropped in a place it's bbox
doesn't fit, it uses the player's bbox (crouched or full size).
This prevents it getting stuck so physics can still move it, and it's
expanded to full size if it's moved to a place where it does fit.

Only adding this to flags in this MR because they already had half the
code. I'll add it to other big ents that players can drop (keys,
balls) in another MR.