Q1BSP: use engine default for gl_max_lightmapsize (currently 512)
This more than doubles average FPS and improves worst-case frametimes on
many Q1BSP maps, and has no effect on Xonotic / Q3BSP maps (as tested
with: weapon and powerup dlights in timedemo, animated realtime world
lights both with and without shadows in static tests).

In old versions of DP (div0-stable), 512 1024 and 4096 all had equally
bad perf. In current DP the perf is far better but only when the right
setting is used.  For most hardware, old and new, 512 seems to be the
sweet spot. Lower settings are somewhat worse, higher settings are MUCH
worse.

Signed-off-by: bones_was_here <bones_was_here@xonotic.au>