Allow g_balance_teams_prevent_imbalance without g_balance_teams
I disagree with the comment because g_balance_teams is a misnomer: all
it does is change whether the team selection UI appears when pressing
+jump. Selecting a team manually can improve balance when players know
each other's skills, and the fact that's in use doesn't automatically
mean that people should be allowed to stack mid-match.

Added a condition to only enforce during the match or countdown, because
it's easier to reorganise teams during warmup when players can directly
switch to their team, instead of one of them needing to spec before the
other can swap.