join queue: UI improvements and refactoring
Fixes duplicate and redundant chatcon notifications.

Displays relevant centreprint each time the team selection GUI or +jump
are used to try to join (even if the player is already queued) for
better noob friendliness.

Always notifies the player when their team selection conflicts with that
of a queued player.

Notifies queued players to get their attention when they join as it may
take some time before another player triggers the join.

Includes the team the player actually got assigned to in the relevant
centreprint (for the cases where they had no preference, or someone else
chose their preferred team first).

This required some refactoring:

Calls queuePlayer() from Join() instead of joinAllowed() which restores
joinAllowed() to only checking and not taking any action for a more
consistent API and better behaviour with sv_spectate 0.

Calls TeamBalance_JoinBestTeam() and SetPlayerTeam() from
PutPlayerInServer() instead of Join(), mostly because it's already done
there for bots which don't always use Join(), and to consolidate the
callsites for these functions.

No longer sets the .team field when adding a player to the queue
(they're not actually on the team yet).

Uses the .team_selected field to pass selection conflicts to Join(), it
can't be done with a parameter because `join` and `selectteam` are
separate commands.

Fixes some uses of NOTIF_ONE where it should have been NOTIF_ONE_ONLY.