add Infection gamemode: survivors must avoid getting infected
The game begins with only survivors. After a few seconds, a random
player will be selected to become infected. If a survivor is killed by
an infected player, the survivor will become infected. If all survivors
are dead, the infected win. If the infected fail to kill all survivors
within the timeframe, survivors win.

This commit is incomplete and will not work. Current issues:
- Fix game crash when a survivor is killed
- Mutators for setting what weapons infected have at their disposal