Improve sv_spectate 0 notifications and description, refactor nospectators command
The blockSpectators global is redundant.
See https://gitlab.com/xonotic/xonotic-data.pk3dir/-/merge_requests/1289

Players are now given the grace period and SPECTATE_WARNING when
sv_spectate is disabled during a game, as well as when `nospectators` is
voted.

Centreprints are added and displayed when they don't conflict with join
restriction centreprints.