Display physics hud's max acceleration properly on config
Previously when configuring the hud, it used an acceleration value
relative to _acceleration_max. This meant players with the acceleration
text off saw no visible change when changing the cvar.
Configuring the hud now uses a static 0.45 acceleration so the change is
reflected. 0.45 was shown before with the default 1.5 _acceleration_max