W-turn indicator fixes
Fixed conditions for displaying the indicators.
Fixed the check for if aiming at the angle would put you in the accel
zone, while side strafing.
Made it blue rather than light-blue, so it's not confused with the angle
indicator, which is also light-blue by default when in an accel zone.