Improve autoswitching options in CTS
Removed cl_cts_noautoswitch and replaced it with cl_autoswitch_cts which has 3 options:
- 0 (default): fallsback to the cl_autoswitch behavior. This may seem unnecessary, but it can be useful on some maps, manually editing the weapon priority list to choose when you autoswitch and when you don't.
- 1: always autoswitch to the new weapon.
- 2: never autoswitch to the new weapon.
These options are of course only used in CTS, in other gamemodes (including Race) the cvar is ignored and cl_autoswitch is used.
Also made the new cvar apply to cases where a trigger gives a player a weapon in CTS, which for some reason it didn't apply to before.
Added a bit of documentation to affected functions.
The new cvar name is easier to find and more intuitive in my opinion.