Allow selecting a random nade
This change allows selecting a random off-hand/bonus nade with cl_nade_type 0, g_nades_nade_type 0, and/or g_nades_bonus_type 0.
The random nade displays as a rainbow on the ammo HUD, drawn by reusing the Normal Nade image.
I didn't create a new nade type, and thus these changes didn't add the option to place random nades on the map, since I didn't feel this would be necessary, and it's ambiguous what random means in that case -- will it be randomly decided each time a player picks it up, or will the type be randomly decided at the start of the round?