Improve hudpanel dialogs a bit
Changed some textsliders to radiobuttons, and added text options to some sliders (by converting them into textsliders).
The implementation of the textsliders needed a weird workaround in this case so that theIdentifier would have the right number of decimal places, or else the default of some of these cvars would be perceived as not in the list, and thus show up as Custom. Perhaps a mixed numbers-text slider is needed in future.