Slightly refactor mutators code
These changes allow mutators to be properly shown in the guide.
Renamed .m_name to .mutatorname, and added (new) .m_name which would be displayed in the guide.
Allowed use of classes other than Mutator in the Mutators registry, so that m_name, m_icon, and m_color can be stored in such class and then displayed in the guide.
Also removed itemstime and waypointsprites mutators from the guide.