Remove redundant team check in sv_tka.qc
`ball` never has a .owner, so this condition should never be true.
This code was likely intended to make bots not chase the ball if their team owns the ball, however this can be perceived as a strategy (protecting the carrier), so I opted to just remove the redundant code and not alter bot behavior.