Make menu_tooltip 2 options look more like the console
Made cvar names be drawn in yellow (^3), and the default value be separated like ["default"] in advanced menu tooltips.
Since the user has to explicitly enable this option, and most people who enable it will almost certainly know about the console, I think this is a worthwhile change to make the two more consistent, given they are ultimately describing the same thing.