Undo physics changes related to hash change of 0e542a1e
This commit altered how physics code works, changing the hash.
This is a regression.
The rest of the changes in 0e542a1e didn't update the hash, and logically speaking shouldn't alter how physics works.
Undid changes which led to the hash change, at the tiny cost of some lost regained performance.