Fix autoswitch regression from !1398
cl_autoswitch_cts 1 always forced switching to the new weapon, but never checked that the weapon was actually given to the player.
Some triggers actually remove a weapon, rather than give it.
This means on some maps like nood-tbb, triggers could set the players held weapon to one they own (on nood-tbb it tries to gives you the Wyvern Attack, previously the Porto-O-Launch I think).
As a side note I'm not really sure why Wyvern Attack is in the Weapons registry, but that's besides the point.
I fixed this by only updating `last_wep` if we actually have the weapon after `GiveWeapon` runs (which may remove the weapon).