Further improve cts autoswitching logic
Previously it switched to the last weapon given or not-taken-away by the trigger, now it switches to the last weapon given by the trigger.
For example on nood-tbb if you give yourself a Vortex then jump on the MachineGun-giving trigger it'll switch to MachineGun rather than staying with Vortex simply because the trigger happened to not remove it (even though you have no Cells).