Match Q3 wishspeed and wishdir calculations
This is equivalent to the relevant change in
0e542a1e3052b8bfe744b6ab34fe62de005b379d and produces the same pipeline
hash (ie the same hash as given by reverting
bec7b1df67bdb738bd65cdb261fe4e6a30890073).

It's written `wishvel * (1 / wishspeed)` because that's how Q3 does it.
The difference in the result is insignificant for normal gametypes but
might be significant for Q3 DeFRaG compatibility in extreme cases.

See https://gitlab.com/xonotic/xonotic-data.pk3dir/-/merge_requests/1438
for discussion.