Allow binding weapons individually in the menu
Previously weapon-groups could only be bound via the menu, meaning a player can't set a separate bind for Crylink and HLAC (for example).
Added individual weapon bind keys under the group key in the menu.
- Binding a group key will unbind all the child weapon binds, so that the user doesn't end up with multiple binds set for the one weapon. This is of course a legitimate thing to do, but is likely too confusing for inexperienced players.
Added tree icons to make it more obvious which weapons are a part of the group.
- Currently there's no tree icons for wickedx or luminos.
Since this now shows weapon icons along with their names, this is enough to complete the menu-side of #2065.