Trim announcer/default/headshot.ogg
Noticed while playing that there was an excessive delay before and after the sound, such that if there were multiple "headshot" sounds in the announcer queue it would take a while for them to finish, and there would be a noticeably long pause between each one.
Trimmed the start and end of the sound, using mp3splt-glk.