Properly disallow selection of disallowed nades
Previously with g_nades_client_select, nade types could be selected with cl_nade_type even if the specific type was disabled, for example g_napalm is 0, but napalm nades could be selected regardless with cl_nade_type 2.
This fixes that, by instead throwing a normal nade if a disallowed nade type is chosen by the player.
This check isn't run on g_nades_nade_type and g_nades_bonus_type, so server admins can have setups such as the following:
- Allow all nades except for ice to be chosen by the player as their offhand nade (since ice nade is a bit overpowered), reserve ice nade as the bonus nade.
- Etc.
Updated the nade type cvar descriptions to reflect that they now need to be set to 1 to allow client selection.
This technically does impact balance, but is mostly a minor change, and is also actually a bug fix, so likely doesn't need a vote.
It may be worthwhile now reconsidering which nades are allowed, currently everything is disallowed except ice, translocate, spawn, heal, and normal.