Make website links in the guide not translatable
Website links shouldn't be translatable, and removing them from the transifex string has the benefit that if the link ever needs to be updated, we won't need every single transifex string to be changed.
Also improved the XonStat guide section a bit, adding the stats.xonotic.org link, and some extra information.