Allow displaying the winning map/gametype on the voting hud
Also renamed g_maplist_votable_nodetail to g_maplist_votable_detail, and added sv_vote_gametype_detail.
When 1, the hud displays the "(2 votes)" text in light blue for all maps in the tied winning place, provided they've received at least one vote.
When 2, the hud displays only which map will win the tie break (randomly chosen by the server) in light blue.
This can help prevent annoying situations where for example 4 players vote for different maps and 1 player votes for a map that the 4 players all don't like, previously there'd be a 20% chance it'd be selected despite the 4 players all disliking it -- now the players will know if it'll be picked or not, if this cvar is set to 2.
This does however remove most of the randomness from the voting, so this is disabled by default (cvars both 1).
Also cleaned up associated code.