join queue: leave queue with F3 (spec) when queued to swap teams
Fixes erroring out the server or killing the player when they're queued
to switch teams and try to leave the queue.

IIRC the initial implementation put the player on the team they wanted
to join even when they were only queued, which would explain the
SetPlayerTeam() call (which causes the server error) in
ClientKill_Now_TeamChange().