join queue: leave queue with F3 (spec) when queued to swap teams
Fixes erroring out the server or killing the player when they're queued
to switch teams and try to leave the queue.

The initial queue implementation put the player on the team they
selected even when they were only queued not joined. That required the
now-redundant SetPlayerTeam() call in ClientKill_Now_TeamChange() (cause
of the server error()) and .wants_join check in
TeamBalance_GetTeamCounts() (also removed here).