Add DPS information to the guide
The guide now has a short sentence about the damage and damage per second (DPS) for each weapon that can deal damage.
Values are the defaults (cvar_defstring).
Strings/Functions may need to be updated if the weapons' functionalities are ever changed in future, but this is unavoidable without an overly complicated solution.
Electro secondary is currently inaccurate since _secondary_refire and _secondary_refire2 are misnamed.