join queue: fix respawn bug when queued to swap teams
Fixes the queued player being swapped when respawning even when there
wasn't another player ready to swap with them, and not being removed
from the queue when swapped to a different team.

Also fixes TeamBalance_JoinBestTeam() being called every time the player
respawned in the campaign, which had no effect but isn't correct.