teamplay: move and rewrite player team forcing code
AFAICT this code is only used for the player component of bot_vs_human.
There's a TODO in TeamBalance_CheckAllowedTeams() about supporting more
than 2 teams with bot_vs_human, in which case there could be more than 2
human teams so the players could switch and/or use the join queue.

Running this code after ClientKill*() code was wrong because the delayed
team switch centreprint could tell the player it was switching them to
team foo, but then they'd be put on team bar. It also needs to run
before the player is put on the join queue, otherwise same bug. This
means moving it to ClientCommand_selectteam().

Moving it there will allow TeamBalance_QueuedPlayersTagIn() to assume
ClientKill_TeamChange() will put the player on the specified team,
allowing simpler code in a following commit.

The code was fragile and verbose so it's replaced with a loop that will
work for any NUM_TEAMS value.