teamplay: when counting team members include in-progress team changes
g_balance_kill_delay causes team changes to be delayed which creates a
window during which balancing code could give incorrect results if this
isn't considered. For example spectator(s) could join (no delay on that)
such that now the team change would stack the teams.

This fixes it such that player(s) will be blocked from joining (what's
about to be) the larger team, and be placed on another team or the join
queue (according to config and whether they chose auto selection).