Allow hiding the ListBox scrollbar
Previously the scrollbar would display even if it couldn't be used, which looked weird in some situations.
Now it only displays if needed, or if explicitly told to.
So far it's only explicitly shown for the side bar in Settings > Game, since it looks jarring to have no vertical separator, which the scrollbar acted as one.
Also fixed a DivZero bug that occured when dragging a non-existent scrollbar.
- This is still important due to the above.