Fix zoom scrolling when spectating
Due to !1411, when you spectated someone and scrolled if they were zooming, it would zoom scroll instead of switching to spectate someone else.
Now it only accounts for whether you're pressing zoom binds yourself to determine whether it should zoom scroll or switch spectators.
This is still a little confusing to use because of cl_unpress_zoom_on_* cvars, but it fixes this bug.