Add sv_airstopaccelerate_full cvar
In Q3DF CPM players' acceleration is multiplied by 2.5 when aiming beyond 90 degrees, provided they're not A/D-strafing (side-strafing).
This didn't happen previously in Xonotic, and instead the multiplication would be sinusoidal, going from 1 to 2.5 over 90 to 180 degrees.
Adjusted code to allow for this by adding sv_airstopaccelerate_full, set it to true in only physicsCPMA.cfg.
Updated StrafeHUD to account for this.
- Previously it also applied sv_airstopaccelerate even when A/D-strafing (when it shouldn't).