Pause warmup countdown when a vote is called
Did this by adding frametime to game_starttime every frame (in VoteThink).
Also added an extra 10 seconds to warmup time when a vote is called, so that people can't abuse this by calling a vote 0.01 seconds before warmup ends, so that (if accepted) there's no way someone else could call a vote afterwards.