Make menu slider animations smooth
Previously the smooth quadratic animation was used only for when the slider was dragged out of range, which didn't work properly since if you dragged it back into the range before the animation was complete then it would bug out.
Made sliders always use the smooth quadratic animation, except in cases like for loadCvars.
Added cvar menu_animations to control how long the animations last, added a slider for this in Settings > Misc.