Allow displaying who used suggestmap
When hovering over a map in the map voting screen, if g_maplist_votable_show_suggester (new cvar) is 1, it'll now display "Suggested by: username".
When voting finishes, if a map that was suggested wins, the text will properly relocate itself.
Made the mapvoting HUD 5% thinner (not in gametype voting) so that the suggester text doesn't cause a 3x3 grid of maps (9) to display differently when abstaining is enabled and the chat takes the full 20% vid_conheight.
Also abstain text now fades out when voting finishes, like map pics and names.
Also removed some useless variables.
Fixed g_maplist_votable_suggestions cvar description, so it reflects the behavior of the code.
Improved documentation of the code.