teamplay: use xonstat skill ratings in team auto-selection
Also fixes team scores being compared in warmup and having too much
influence when it's too early for them to be reliable.

Behaves differently in warmup (see comments and the sv_teamnagger
cvar description) to prevent a common problem with always selecting the
weakest team.

Thanks k9 for suggestions to use Cohen's d and precision weighting.