teamplay: use xonstat skill ratings in team auto selection
Also fixes team scores being compared in warmup and having too much
influence when it's too early for them to be reliable.

Removes the `use_score` parameter as it's too coarse and the only
callsite where it was disabled should have enabled it to prevent
switching to the winning team when the losing team is larger by 1.