minelayer: Delay remote detonation until teammates are far
Previously mine remote detonation didn't occur if the mine's owner was within the remote_radius, but it didn't care if the owner's teammates were in within the radius.
Made it delay detonation, making it consistent with the rest of the Mine Layer's code regarding teammate-sparing.
Also added a WEAPONTODO comment about the protection cvar.