menu: adjust logic and appearance for full servers
Comments out the nJoinAllowed() slot count filtering because there's
several reasons for connecting to servers where that func reported
0 free slots (see comments).

Applies the EMPTY alpha value to servers where nJoinAllowed() reported
0 free slots (not highlighted like a server reporting free slots, but
also not dimmed like one that can't be connected to).

Applies the FULL alpha value to favourite servers that can't be
connected to (if they aren't filtered out).

Removes the negative ping filter because that bug was fixed in
https://gitlab.com/xonotic/darkplaces/-/commit/fe2b222e32752c408fbe5ccf4ac3fb546f8ff41a

Makes a few clarifications.