ctf: fix flags occasionally getting thrown out of the world on q3 maps
The flag bbox was placed below the player, often mostly in the ground,
at the start of the trace. When sv_q3compat_changehitbox is enabled the
player's mins.z is 4u smaller on q3 maps so it was placed even further
in the ground. Sometimes nearby geometry prevented the trace from moving
it up and out. Sometimes the floor brushes are quite thin so the bbox
was mostly outside the world.

This created a worst-case scenario for nudgeoutofsolid() so it
occasionally unstuck the flag by nudging it the rest of the way out of
the world. It also made nudgeoutofsolid() more likely to fail, requiring
the fallback to be used (not ideal for gameplay).