teamplay: improve UX for trying to join a disallowed team
When a player selects a specific team that they're explicitly disallowed
from joining, this informs them of that instead of "silently" putting
them on a different team.

Removes the autocvar_bot_vs_human check and uses a generic message in
case this team allowed code is used for something else in future.

Reintroduces the same notification identifier that was used for
g_changeteam_banned (removed) and slightly optimises the code.